fix: some tests
fix: pmd complaint

fix: missing restore

fix: set some strings as final

tidy up

tidy up 2

tidy up 3

fix: make duration type and timeslot strings global constants

fix: undo order chane in hearingdetails

fix: changes for lucas comments

fix: make too small session duration simpler

fix: some tests

fix: pmd complaint

fix: missing restore

fix: set some strings as final

tidy up

tidy up 2

tidy up 3

fix: make duration type and timeslot strings global constants

fix: undo order chane in hearingdetails

fix: changes for lucas comments
