+++ src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/HearingDetails.java
@@ -24,6 +24,10 @@ public class HearingDetails {
 
     private HearingWindow hearingWindow;
 
+    private Number duration;
+
+    private List<String> nonStandardHearingDurationReasons;
+
     private String hearingPriorityType;
 
     private Number numberOfPhysicalAttendees;
