--- src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/HearingDuration.java
@@ -1,21 +1,0 @@
-package uk.gov.hmcts.reform.sscs.model.single.hearing;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class HearingDuration {
-    private Number duration;
-    private List<String> nonStandardHearingDurationReasons;
-    private Integer durationBenefitIssueCodes;
-    private Integer adjournmentDuration;
-}
