+++ src/test/java/uk/gov/hmcts/reform/sscs/helper/mapping/HearingsDetailsMappingTest.java
@@ -19,7 +19,6 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.CcdValue;
 import uk.gov.hmcts.reform.sscs.ccd.domain.ElementDisputed;
 import uk.gov.hmcts.reform.sscs.ccd.domain.ElementDisputedDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingOptions;
-import uk.gov.hmcts.reform.sscs.ccd.domain.HearingSubtype;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Issue;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Name;
 import uk.gov.hmcts.reform.sscs.ccd.domain.OtherParty;
@@ -54,8 +53,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 import static uk.gov.hmcts.reform.sscs.model.hmc.reference.LocationType.COURT;
@@ -82,7 +79,7 @@ class HearingsDetailsMappingTest extends HearingsMappingBase {
         given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
             .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
                                             60, 75, 30
-            ));
+           ));
         given(sessionCategoryMaps.getSessionCategory(BENEFIT_CODE, ISSUE_CODE, false, false))
             .willReturn(new SessionCategoryMap(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
                                                false, false, SessionCategory.CATEGORY_03, null
@@ -538,283 +535,6 @@ class HearingsDetailsMappingTest extends HearingsMappingBase {
         assertNull(result);
     }
 
-    @DisplayName("when a invalid adjournCaseDuration or adjournCaseDurationUnits is given getHearingDuration returns the default duration Parameterized Tests")
-    @ParameterizedTest
-    @CsvSource(value = {
-        "null,null",
-        "null,60",
-        "1,test",
-    }, nullValues = {"null"})
-    void getHearingDuration(String adjournCaseDuration, String adjournCaseDurationUnits) {
-        // TODO Finish Test when method done
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                                            60, 75, 30
-            ));
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .adjournCaseNextHearingListingDuration(adjournCaseDuration)
-            .adjournCaseNextHearingListingDurationUnits(adjournCaseDurationUnits)
-            .appeal(Appeal.builder()
-                        .hearingOptions(HearingOptions.builder().build())
-                        .build())
-            .build();
-        int result = HearingsDetailsMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-        assertEquals(30, result);
-    }
-
-    @DisplayName("when a valid adjournCaseDuration and adjournCaseDurationUnits is given getHearingDuration returns the correct duration Parameterized Tests")
-    @ParameterizedTest
-    @CsvSource(value = {
-        "2,hours,120",
-        "1,sessions,165",
-    }, nullValues = {"null"})
-    void getHearingDuration(String adjournCaseDuration, String adjournCaseDurationUnits, int expected) {
-        // TODO Finish Test when method done
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .adjournCaseNextHearingListingDuration(adjournCaseDuration)
-            .adjournCaseNextHearingListingDurationUnits(adjournCaseDurationUnits)
-            .appeal(Appeal.builder()
-                        .hearingOptions(HearingOptions.builder().build())
-                        .build())
-            .build();
-        int result = HearingsDetailsMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-        assertEquals(expected, result);
-    }
-
-    @DisplayName("When an invalid adjournCaseDuration and adjournCaseDurationUnits is given and overrideDuration "
-        + "is not present then override the duration of hearing")
-    @ParameterizedTest
-    @CsvSource(value = {
-        "null,75",
-        "0,75",
-        "-1, 75"
-    }, nullValues = {"null"})
-    void getHearingDurationWillNotReturnOverrideDurationWhenPresent(Integer overrideDuration, int expectedResult) {
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                                            60, 75, 30
-            ));
-
-        given(hearingDurations.addExtraTimeIfNeeded(eq(60), eq(BenefitCode.PIP_NEW_CLAIM), eq(Issue.DD), any()))
-            .willReturn(75);
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .adjournCaseNextHearingListingDuration(null)
-            .adjournCaseNextHearingListingDurationUnits(null)
-            .appeal(Appeal.builder()
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend("Yes")
-                    .build())
-                .build())
-            .schedulingAndListingFields(SchedulingAndListingFields.builder()
-                .overrideFields(OverrideFields.builder()
-                    .duration(overrideDuration)
-                    .build())
-                .build())
-            .build();
-
-        int result = HearingsDetailsMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-        assertThat(result).isEqualTo(expectedResult);
-    }
-
-    @DisplayName("When an invalid adjournCaseDuration and adjournCaseDurationUnits is given and overrideDuration "
-        + "is present then override the duration of hearing")
-    @Test
-    void getHearingDurationWillReturnOverrideDurationWhenPresent() {
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .adjournCaseNextHearingListingDuration(null)
-            .adjournCaseNextHearingListingDurationUnits(null)
-            .appeal(Appeal.builder()
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend("Yes")
-                    .build())
-                .build())
-            .schedulingAndListingFields(SchedulingAndListingFields.builder()
-                .overrideFields(OverrideFields.builder()
-                    .duration(60)
-                    .build())
-                .build())
-            .build();
-
-        int result = HearingsDetailsMapping.getHearingDuration(caseData, referenceDataServiceHolder);
-
-        assertThat(result).isEqualTo(60);
-    }
-
-    @DisplayName("When the benefit or issue code is null getHearingDurationBenefitIssueCodes returns null Parameterized Tests")
-    @ParameterizedTest
-    @CsvSource(value = {
-        "null,null",
-        "002,null",
-        "null,DD",
-    }, nullValues = {"null"})
-    void getHearingDurationBenefitIssueCodesPaper(String benefitCode, String issueCode) {
-
-        given(hearingDurations.getHearingDuration(benefitCode, issueCode)).willReturn(null);
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(benefitCode)
-            .issueCode(issueCode)
-            .appeal(Appeal.builder()
-                        .hearingSubtype(HearingSubtype.builder().build())
-                        .hearingOptions(HearingOptions.builder().build())
-                        .build())
-            .build();
-
-        Integer result = HearingsDetailsMapping.getHearingDurationBenefitIssueCodes(
-            caseData,
-            referenceDataServiceHolder
-        );
-
-        assertThat(result).isNull();
-    }
-
-    @DisplayName("When wantsToAttend for the Appeal is null and the hearing type is paper "
-        + "getHearingDurationBenefitIssueCodes return the correct paper durations")
-    @Test
-    void getHearingDurationBenefitIssueCodesPaper() {
-
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                                            60, 75, 30
-            ));
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .appeal(Appeal.builder()
-                .hearingSubtype(HearingSubtype.builder().build())
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend("No")
-                    .build())
-                .build())
-            .build();
-
-        Integer result = HearingsDetailsMapping.getHearingDurationBenefitIssueCodes(
-            caseData,
-            referenceDataServiceHolder
-        );
-
-        assertThat(result).isEqualTo(30);
-    }
-
-    @DisplayName("When wantsToAttend for the Appeal is Yes and languageInterpreter is null "
-        + "getHearingDurationBenefitIssueCodes return the correct face to face durations")
-    @Test
-    void getHearingDurationBenefitIssueCodesFaceToFace() {
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                                            60, 75, 30
-            ));
-
-        given(hearingDurations.addExtraTimeIfNeeded(eq(60), eq(BenefitCode.PIP_NEW_CLAIM), eq(Issue.DD), any()))
-            .willReturn(60);
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .appeal(Appeal.builder()
-                        .hearingOptions(HearingOptions.builder()
-                                            .wantsToAttend("Yes")
-                                            .build())
-                        .build())
-            .build();
-
-        Integer result = HearingsDetailsMapping.getHearingDurationBenefitIssueCodes(
-            caseData,
-            referenceDataServiceHolder
-        );
-
-        assertThat(result).isEqualTo(60);
-    }
-
-    @DisplayName("When wantsToAttend for the Appeal is Yes "
-        + "getHearingDurationBenefitIssueCodes return the correct interpreter durations")
-    @Test
-    void getHearingDurationBenefitIssueCodesInterpreter() {
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                                            60, 75, 30
-            ));
-
-        given(hearingDurations.addExtraTimeIfNeeded(eq(75), eq(BenefitCode.PIP_NEW_CLAIM), eq(Issue.DD), any()))
-            .willReturn(75);
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .appeal(Appeal.builder()
-                        .hearingOptions(HearingOptions.builder()
-                                            .wantsToAttend("Yes")
-                                            .languageInterpreter("Yes")
-                                            .build())
-                        .build())
-            .build();
-
-        Integer result = HearingsDetailsMapping.getHearingDurationBenefitIssueCodes(
-            caseData,
-            referenceDataServiceHolder
-        );
-
-        assertThat(result).isEqualTo(75);
-    }
-
-    @DisplayName("When wantsToAttend for the Appeal is No and the hearing type is not paper "
-        + "getHearingDurationBenefitIssueCodes return the correct paper durations")
-    @Test
-    void getHearingDurationBenefitIssueCodesNotAttendNotPaper() {
-
-        given(hearingDurations.getHearingDuration(BENEFIT_CODE, ISSUE_CODE))
-            .willReturn(new HearingDuration(BenefitCode.PIP_NEW_CLAIM, Issue.DD,
-                60, 75, 30
-            ));
-
-        given(referenceDataServiceHolder.getHearingDurations()).willReturn(hearingDurations);
-
-        SscsCaseData caseData = SscsCaseData.builder()
-            .benefitCode(BENEFIT_CODE)
-            .issueCode(ISSUE_CODE)
-            .appeal(Appeal.builder()
-                .hearingSubtype(HearingSubtype.builder().build())
-                .hearingOptions(HearingOptions.builder()
-                    .wantsToAttend("No")
-                    .build())
-                .build())
-            .dwpIsOfficerAttending("Yes")
-            .build();
-
-        Integer result = HearingsDetailsMapping.getHearingDurationBenefitIssueCodes(
-            caseData,
-            referenceDataServiceHolder
-        );
-
-        assertThat(result).isNull();
-    }
 
 
     @DisplayName("getElementsDisputed returns empty list when elementDisputed is Null")
